parser/requirements.txt
telethon>=1.34.0
aiokafka>=0.10.0
python-dotenv
minio>=7.1.14 